Compile and link with 'getopt.c' explicitly.
This should not be necessary, but for some reason
(that completely escapes me!), using command-line options
(such as -IR) with the Release version of snmptranslate
causes a run-time DLL error.  (The Debug version is fine).
The other applications seem to be OK - this problem only
seems to hit snmptranslate.  So as a sticking-plaster,
compiliing getopt.c in as part of the project, keeps both
versions of the application happy.
  Someone with better Windows knowledge than me needs to
look at this and fix it properly.


git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@7621 06827809-a52a-0410-b366-d66718629ded