(excellent) patch #705566 from Andrew Findlay on perl agent support:
1) Agent cannot return MAC addresses containing zero octet.

2) Agent coredumps when trying to return IP addresses
if these are supplied as dotted-quad or as a hostname
that fails to resolve.

3) Agent corrupts OID values.

4) Agent truncates any OCTET_STRING or BIT_STRING that
contains a zero octet (this is the root cause of (1) above)

API changes introduced:

a) IP addresses now passed as binary strings (4 octets
for IPv4) rather than dotted-quad or name. This
sidesteps the IPv4/IPv6 problem and leaves the calling
script to deal with name->address resolution and
associated error handling.


git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@8348 06827809-a52a-0410-b366-d66718629ded