I've never like the renaming of the container-based table_data API calls and data structures.  If we're starting to think about releasing this code, then any alterations must be made *before* this goes live, or else it'll be too late.
I've finally come up with a form of naming for the table_data2
code that I'm reasonably happy with (based around 'tdata'),
so this patch rewrites this helper to use that style (as well
as tweaking the dataset2 code to match).
  The main dataset2 API hasn't been properly updated, but this
should probably also be addressed before 5.3 goes out of the door.

This version also illustrates a suitable "generic table API",
that could usefully be considered for other table helpers as
well.  Some aspects of it may be specific to the table_data
framework, but the basic ideas are more widely relevant.


git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@12780 06827809-a52a-0410-b366-d66718629ded