- a tangled web of fixes for override   - back out fix for 711465: override directive ignored with snmpbulkget   - apply new fixes for 711465 (see bug report for gory details)   - also fix override of table instance, reported on coders - and, quoting Dave:
I'm not at all sure that this is necessarily the correct fix, but
it wasn't very clear as to exactly what is being handled here.
This change certainly addresses the above bug report, without
seeming to break anything else.

Let's give this patch time to settle into the main development code,
and then consider back-porting it to earlier lines.....


git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@14243 06827809-a52a-0410-b366-d66718629ded