CHANGES: snmpd: fix crash in IF-MIB and IP-MIB when there are some interfaces with broadcast IP address and some without.
Reset the bcastentry in the loop, so it cannot be inserted to the container
twice.

Without this patch, when an interface without broadcast address is processed,
the bcastaddress from previous loop is inserted into the container. This
obviously fails and the bcastentry is freed, but the container still contains
reference to it -> SIGSEGV.

git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@17931 06827809-a52a-0410-b366-d66718629ded