agent: change type of snmp_enableauthentraps to long
Type of snmp_enableauthentraps variable is changed from int to long.
This fixes the bug with writing to snmpEnableAuthenTraps if server is
compiled for 64bit (snmpset reports wrong length). This bug was
introduced a long time ago (by faeecd0 commit).

Actually this is a 'less effort' type of fix but I wonder why we need a
long to store an INTEGER in the first place?

Signed-off-by: Ilya Yanok <yanok@emcraft.com>
Signed-off-by: Wes Hardaker <hardaker@users.sourceforge.net>