CHANGES: Windows: perl: Make sure that the constant SNMP_DEFAULT_RETRIES is recognized.
CHANGES: perl: SNMP module: Add constants NETSNMP_CALLBACK_OP_RECEIVED_MESSAGE and NETSNMP_CALLBACK_OP_TIMED_OUT.

When the Net-SNMP Perl modules are built on Windows with MSVC or MinGW it
is possible that each these are linked with another version of the MSVC
runtime library than the Perl interpreter itself. In that case the "errno"
variable used inside .xs source files is another variable than the $! variable
in .pm files. Or: assignments to "errno" in a .xs file do not modify the "$!"
Perl variable This patch avoids using "errno" by modifying the different
constant() functions such that these return a two-element array instead of
returning a scalar and setting errno.