CHANGES: building: Silence libbsd 0.3.0 nagging
If you have libbsd 0.3.0 installed then their version of nlist.h warns and
tries to make you include bsd/nlist.h instead. The problem is that we don't
need that compatibility library as we know how to do it ourself so it becomes
pure nagging. This patch adds a define that changes the warning into an
error and then figures out that there is no usable nlist.h and ignores it
for the future.