CHANGES: snmplib: Set the rpath for libperl.so in libnetsnmpmibs.so such that the dynamic linker can find libperl.so without requiring the user to modify the dynamic linker settings.
This change in agent/Makefile.in is necessary on the 5.5 branch and
later but not on the 5.4 branch. That is because the 5.4 branch uses
an older version of libtool that propagates the rpath of dependent
libraries. This is called inter-library dependency tracking in the
libtool manual. This functionality is present in libtool version 1
but not in libtool version 2. For more background information about
dynamic libraries and rpaths, see also the document with the title
"RpathIssue" on the Debian wiki (http://wiki.debian.org/RpathIssue).