CHANGES: snmp-bridge-mib: Fix index interpretation.
This patch corrects a wrong data interpretation.
snmp-bridge-mib obtains the ifindex value from the sysfs
attribute 'ifindex' The value given by the sysfs attribute
is an integer, which is handled as hex and leads to
incorrect data displayed to the user.

e.g.
    $ > snmpwalk localhost BRIDGE-MIB::dot1dBasePortIfIndex.2
    BRIDGE-MIB::dot1dBasePortIfIndex.2 = INTEGER: 54

according to the sysfs attribute the value is 36

    $ > cat /sys/class/net/<device>/ifindex
    36

Signed-off-by: Mijo Safradin <safradin@linux.vnet.ibm.com>