CHANGES: snmplib: Be more robust if clock_gettime fails.
On RHEL3 it seems clock_gettime(CLOCK_MONOTONIC, ) always returns EINVAL but
it works for the other clocks.

Further, the old code which returned a constant { 0, 0 } caused the test suite
to get stuck in a loop if no response arrived - thus this code have the
advantage of working less bad.