CHANGES: snmplib: Rename memdup() into netsnmp_memdup()
Rename the function memdup() defined in libnetsnmp.so into
netsnmp_memdup(). This avoids conflicts with memdup() functions
defined in other libraries that define a function with the same
name. As an example, if WINS name resolution has been enabled via
/etc/nsswitch.conf then performing a host name lookup can trigger
a crash in the SAMBA WINS name lookup library because calls to
memdup() from that library get redirected to the incompatible
Net-SNMP memdup() function. See also
https://bugzilla.redhat.com/show_bug.cgi?id=1000780 for the details.
This patch intentionally changes the Net-SNMP ABI.