CHANGES: snmplib: Avoid failure to trigger sign extension due to overlong types
In asn_realloc_rbuild_signed_int64 the algorithm assumes that shifting
values greater than 0x7fffffff right is an arithmetic shift that will
shift in 1's. On hosts where long is more than 32 bits the use of long
for 'low' and 'high' makes this assumption false.

Thanks to rkj@softhome.net for reporting this issue.