Proxy to self and make sure walk works.
This test is for the case described in
https://sourceforge.net/p/net-snmp/bugs/2578/
where the reporter upgraded from 5.3.1 to 5.7.2.1
and started getting "Error: OID not increasing"
when walking a proxied tree.  I used the system
tree instead of the reporter's enterprise tree,
but otherwise believe that the replication is
correct.  This test passes under 5.7.3, so I've
passed it back to the reporter to try 5.7.3 before
we dig deeper.