CHANGES: snmplib: Fixed reporting 'error writing to /var/xxx/snmpapp.conf'.
When a client utility, such as snmptrap, tries to write to its persistent
configuration file (/var/net-snmp/snmpapp.conf in Fedora), do not report
any error when open() fails. The tool is typically run by non-root, who
cannot write to /var and the error just confuses users.

And when doing it, make sure that "snmpapp" string is defined only on one
place, just in case.