getbulk enhancements: response size + fallback to forward encoding
- only check reverse encode default store flag once per packet, setting
pdu msg flag for forward encoding if reverse encoding is not
configured.
- track rough response sizes as we go, so we can stop gathering response
varbinds we probably won't be able to send
- if initial reverse encoding results in a packet that is too big, set
pdu flags to retry with forward encoding and drop varbinds so that
response fits in packet size
- new private asn type to indicate encoding processing should stop
- move duplicated packet building code to new netsnmp_build_packet()
- new buffer ptr/len in agent isp struct for outgoing packet
- new _build_initial_pdu_packet() to build a response packet (saving
result in new isp struct ptr/len). this function is called by the
agent to build the response *before* snmp_send is called, so we can
a) detect tooBig errors and b) so we can fix getbulk processing by
detecting bulk responses that exceed max snd msg size and switching
to forward encoding with varbind truncation.
- add ':intermediate' to intermediate results debug tokens

Signed-off-by: Robert Story <rstory@freesnmp.com>