Fixed potential bad free.
When a 'default_domain' starts with a comma ',', the first strtok()
returns pointer somwehere to the 'default_domain' duplicate and not
to the first character. When cleaning up, make sure we try to free
the duplicate (i.e. tokenized_domain) and not the first found token
(which may be different).