Revert "Trim output of net-snmp-config --cflags."
This reverts commit 127274f8f27fba264a4c0f91190fd6f8037cdf21. Robert Story
explained as follows on the net-snmp-cvs mailing list why this patch is
wrong and needs to be reverted:
<quote>
Actually, I --cflags is supposed to be everything used to compile
the agent, so that subagents and modules for dlmod loading can be
built exactly the same.

In particular, removing other flags can remove options that will
change structure sizes when parsing the headers, resulting in
binary incompatibilities.

If you only need the include directories for some reason, please
add another option.
</quote>