If the status goes disabled after it starts, the probe is completed.
As the comment said: if you are pinging something that does not
respond at all, the responses will never increase, and we should
be querying pingResultsSentProbes to make sure that the probe has
started.  Now, if pingResultsSentProbes is nonzero and the status
becomes disabled, we consider the test to be completed.