Fix/enforce ipDefaultRouterLifetime value range
IP-MIB defines ipDefaultRouterLifetime to be in 0..65535 range, but the
current code 1) erroneously uses 0xFFFFFFFFUL (4294967296) as the max
value, and 2) does not clip the computed value that exceeds the maximum
(this was unnecessary probably because the max value was also the max
value permitted for the underlying type uint32_t).

Fix IPDEFAULTROUTERLIFETIME_MAX to be 65535 (0xFFFF) and clip
out-of-range values to the max value.