NEWS: snmplib: PATCH: 1349: Fix perl/other crash against bad SNMPv3 agent
With the patch in 1214, the snmp_api code assumed that if magic was
set, it was the "struct synch-state" from snmp_client.  Of course,
magic belongs to the caller, and the perl library uses it differently,
so reaching into it is verboten.  Introduce a new callback (that
was already introduced in 5.8) to report this "retries exceeded"
state, and use it in snmp_client.