Add Travis and Appveyor CI support
This patch makes it possible to enable Travis and Appveyor continuous
integration support. On Travis Net-SNMP is built with all MIBs, with
--disable-set-support, with --enable-mini-agent, with --enable-minimalist
and with --enable-read-only. All these configurations are built on Linux
and on OS/X. On Appveyor Net-SNMP is built with MSVC and dynamic
libraries, with MSVC and statically linked libraryes, with MinGW64, with
Cygwin32 and with Cygwin64. As one can see in the scripts in the "ci"
directory the to-do items are as follows:

- Fix the TLSTCP test on Travis.
- Fix the MinGW64 build. That build fails with a linker error.
- Ensure that walking the entire MIB works on Darwin.
- Make the regression tests pass for the Cygwin builds.
- Make the regression tests pass for --disable-set-support,s
  --enable-mini-agent, --enable-minimalist and --enable-read-only.
- Make the non-default MIBs compatible with --disable-set-support,
  --enable-mini-agent, --enable-minimalist and --enable-read-only.