snmplib: fix clientaddr
cea974008da30de8d39a1f8ede7189265d93d598 ("clear stack variable before
use") zeroed out the addr_pair, which was cleared at the beginning
of the function and then the remote address was stored in it.
Clearing it here results in clearing the remote address, meaning
that if you create a session with a clientaddr, the remote address
gets changed to 0.0.0.0.  Deleting this line restores the correct
behavior of sending to the desired remote address, and since the
memset() at the beginning of the function already cleared the stack
variable, it added no benefit.