libsnmp/transports/UDP: Remove MSG_NOSIGNAL
Remove MSG_NOSIGNAL from the UDP code because it does not affect SOCK_DGRAM
sockets. From the sendto() specification
(http://pubs.opengroup.org/onlinepubs/9699919799/functions/sendto.html):
"[EPIPE] The socket is shut down for writing, or the socket is
connection-mode and is no longer connected. In the latter case, and if
the socket is of type SOCK_STREAM or SOCK_SEQPACKET and the MSG_NOSIGNAL
flag is not set, the SIGPIPE signal is generated to the calling thread."