libsnmp, ASN parsing: Avoid triggering undefined shift-left behavior
A quote from https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=7199:

So this is a fun corner case of asn.1 integer parsing: the C standard says
that shifting a negative number left has undefined behavior. The other
obvious way to do this is to use the same code on an unsigned long, and
then cast to long.  The result of a cast to a signed value is
implementation-defined.  Linus Torvalds' argument on this is that there
is no sensible way for the implementation to define it other than "copy
the 2's complement bits".

Avoid triggering all these corner cases by using byte-copying instead of
using shift operations.

Fixes: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=7199