libsnmp: Set Linux VRF iface on Trap sink IP addresses
For Linux VRF support in Traps, this patch fixes an issue (#12) where
configuring and iface on the sink IP address was not being bound (via
SO_BINDTODEVICE) to the socket.  This was broken for versions 1, 2c, and 3.

While a clientaddr setting in snmp.conf could fix this, a user would have to
find and set a source IP address in the VRF to be used. This patch does not
require setting a clientaddr but simply binds the iface provided in the sink
IP address (target) to the session socket that will be used.

Signed-off-by: Sam Tannous <stannous@cumulusnetworks.com>
[ bvanassche: made minor edits to patch description and removed one space from
  the patch itself ]