Reduce CPU load for IF-MIB on Linux servers with large number of interfaces
This patch introduces a new configution option to limit the number
of interfaces that the IF-MIB will process.  On Linux servers with a
large number of interfaces (ppp, dummy, bridge, etc.), the IF-MIB
timer based stats collector can take a large amount of CPU processing.

The new config option "include_ifmib_iface_prefix" takes a space
separated string of ifname prefixes ("eth lo bridg") and will only
include interfaces with these prefixes in the IF-MIB tables.

If this config option is not present (the default), all interfaces
will continue to be processed in the IF-MIB.

Signed-off-by: Sam Tannous <stannous@cumulusnetworks.com>
[ bvanassche: Moved the netsnmp_access_interface_init() call from
  init_mib_modules() into the MIB implementations that depend on that call. ]