Add new config option to limit max number of interfaces for IF-MIB
This patch introduces a new config option "ifmib_max_num_ifaces NUM"
This patch sets the maximum number of interfaces included in IF-MIB
data collection.  For servers with a large number of interfaces
(ppp, dummy, bridge, etc) the IF-MIB processing will take a large
chunk of CPU for ioctl calls (on Linux). Setting a reasonable
maximum number of ifaces for the CPU used will reduce the CPU
load for IF-MIB processing.  For example, configuring
"ifmib_max_num_ifaces 500" will include only the first 500
interfaces based on ifindex and ignore all others for IF-MIB
processing.

Signed-off-by: Sam Tannous <stannous@cumulusnetworks.com>
[ bvanassche: combined two patches into one and removed one unused variable ]