CHANGES: configure: Assume that memcmp() works
From the autoconf documentation: "Macro: AC_FUNC_MEMCMP. If the memcmp
function is not available, or does not work on 8-bit data (like the one
on SunOS 4.1.3), or fails when comparing 16 bytes or more and with at
least one buffer not starting on a 4-byte boundary (such as the one on
NeXT x86 OpenStep), require an AC_LIBOBJ replacement for `memcmp`.

This macro is obsolescent, as current systems have a working memcmp.
New programs need not use this macro."

Hence remove the AC_FUNC_MEMCMP invocation.

See also https://github.com/net-snmp/net-snmp/issues/81.