snmpd: Do not open /dev/kmem on Linux
The Net-SNMP agent only uses /dev/kmem to access kernel data structures on
BSD systems. Hence do not open /dev/kmem on Linux. A side effect of this
patch is that it enables cross-compilation from Linux to MinGW. /dev/kmem
exists on Linux but the code in agent/kernel.c that is surrounded with
 #ifdef HAVE_KMEM / #endif does not build with the MinGW compiler.