perl/SNMP/t/bulkwalk.t: Improve the reliability of this test
Use ok($a, $b) instead of ok($a eq $b) to make error reporting more
informative. Unbless $list[3] before iterating over it to make this
test passs on DragonFly BSD. In the loop that looks for the loop
network adapter, check whether the index of that adapter is numeric
instead of comparing it with one.