libsnmp: Change several constants from unsigned char into unsigned int
The C language requires that the operands of a bitwise or expression
are promoted before the bitwise OR happens. From the C standard:
"The following may be used in an expression wherever an int or unsigned int
may be used:
- An object or expression with an integer type (other than int or unsigned
  int) whose integer conversion rank is less than or equal to the rank of
  int and unsigned int.
- A bit-field of type _Bool , int, signed int, or unsigned int.

If an int can represent all values of the original type (as restricted by
the width, for a bit-field), the value is converted to an int; otherwise,
it is converted to an unsigned int. These are called the integer promotions.
All other types are unchanged by the integer promotions."

In other words, the result of a bitwise OR of two unsigned char operands
has type 'int' instead of unsigned int. Hence change the type of constants
in header files from unsigned char into unsigned int.

See also https://github.com/net-snmp/net-snmp/issues/142.