Read intermediate certificates from PEM files
This patch lifts the restriction that a PEM file can only contain
one certificate. This brings net-snmp in line with the behaviour
of other applications that use PEM files.

Multiple certificates are read from each file, and the offset of
each certificate is added to the index for efficient retrieval.

The first certificate in the chain is eligible to be either an
identity certificate or a remote certificate. Intermediate
certificates are ignored for this purpose, maintaining existing
behaviour with respect to peerCert and remoteCert.

Trusted CA certificates are still required to be listed one per
file, this restriction will be lifted in a separate patch.