Perl, agent: Add precision to double to integer conversion errors
When there's a conversion error due to accumulated rounding errors,
the log messages can very easily omit the relevant non-zero fraction
parts, which makes the warning very confusing, as it appears as if
everything is fine. Increase the precision to 20 characters, which
should be enough for double types, and switch from %f to %g format
conversion which will omit zeros if unnecessary.