libsnmp: Fix the linked list _get() function
If looking for an exact match in a sorted list, return NULL instead of the
next element if there is no exact match. Additionally, inverse the polarity
of the third argument of this function.

Note: I'm not aware of any code inside the Net-SNMP project that is affected
by this change.