Fix deprecated retrans_time syscall warning (and base_reachable_time)
Removed check which retrans_time and base_reachable_time interface was available from  `netsnmp_arch_interface_init`, which uses the "default" network interface to determine which is available, but this network interface is not always available (e.g. network namespaces). 

This fix should on demand determine the available interfaces which are actually available.