snmpd: always exit after displaying usage
Currently, viewing the help text with -h results in snmpd being started
in the background, whereas this does not happen with --help. Similarly,
when an error is detected in command line syntax, the help text is
displayed but sometimes snmpd gets started anyway, depending on the
execution path.

This patch makes snmpd consistently terminate whenever the usage
function gets called. It also removes the goto statements no longer
needed.