ci/net-snmp-configure: Refine when to enable embedded Perl
Only enable embedded Perl if Net-SNMP and the Perl interpreter are built with
the same compiler type. If different compilers are used and if embedded Perl
is enabled, the configure script reports the following:

configure: error: This build is using a GNU C compiler (clang) while Perl
has been compiled with a non-GNU (or non-working) compiler
(x86_64-linux-gnu-gcc). This likely won't work for building with Perl
support. Either specify a different compiler (--with-cc=PATH), disable this
check (--disable-perl-cc-checks) or build without Perl
(--without-perl-modules).