SNMP-NOTIFICATION-MIB: Fix filter lookup
The notification filter information is stored in one container but during
filtering, a different container is being used to look for matching filters.
Change agent/mibgroup/snmp-notification-mib/snmpNotifyFilterTable/snmpNotifyFilterTable_data_storage.h
and agent/mibgroup/snmp-notification-mib/snmpNotifyFilterTable/snmpNotifyFilterTable_data_access.c
to expose the data_storage method that creates the container (used when
actual filtering) so that this method can be used when creating the
container used to store the notification filters.

Fixes: https://github.com/net-snmp/net-snmp/issues/342