configure: Use pkg-config to detect the OpenSSL dependencies
Which libraries must be linked in addition to the ssl library depends on the
configuration of OpenSSL and on the platform. Hence use pkg-config to detect
these libraries if pkg-config is available. Provide a fallback if pkg-config
is not available. See also "compile fails with net-snmp-5.8 and openssl-1.1.1d"
(https://github.com/net-snmp/net-snmp/issues/25).