Solaris: Fix ifHighSpeed
The ifHighSpeed value is not being fetched correctly for Solaris systems as
seen in the following example. For a 10Gbps interface, we see

> snmpwalk -v 2c -c public localhost ifdescr
IF-MIB::ifDescr.2 = STRING: net0
> snmpwalk -v 2c -c public localhost ifspeed
IF-MIB::ifSpeed.2 = Gauge32: 1410065408
> snmpwalk -v 2c -c public localhost ifHighSpeed
IF-MIB::ifHighSpeed.1 = Gauge32: 127
IF-MIB::ifHighSpeed.2 = Gauge32: 1410
> dladm show-phys -Z
LINK            ZONE      MEDIA         STATE      SPEED  DUPLEX    DEVICE
net0            global    Ethernet      up         10000  full      ixgbe0
> kstat -p|grep speed
....
ixgbe:0:phys:ifspeed    10000000000   //10Gbps
..

The expected output is  10000000000 expressed in Mbps, i.e 10000
IF-MIB::ifSpeed.2 = Gauge32: 10000

The issue occurs as we are using an integer(Gauge) type to fetch the ifspeed
value and simply convert it into Mbps for the ifHighSpeed value, which will
fail for speeds > INT_MAX-1.

From the net-snmp documentation at
http://www.net-snmp.org/docs/mibs/interfaces.html: "An estimate of the
interface's current bandwidth in bits per second.  For interfaces which do
not vary in bandwidth or for those where no accurate estimation can be made,
this object should contain the nominal bandwidth. If the bandwidth of the
interface is greater than the maximum value reportable by this object then
this object should report its maximum value (4,294,967,295) and ifHighSpeed
must be used to report the interace's speed.  For a sub-layer which has no
concept of bandwidth, this object should be zero."

The fix will use a uint64 type for storing the ifspeed value, and will
display the ifspeed value as per the docs.

Here's the output after applying the fix
> snmpwalk -c public -v 2c localhost ifSpeed
IF-MIB::ifSpeed.2 = Gauge32: 4294967295
> snmpwalk -c public -v 2c localhost ifHighSpeed
IF-MIB::ifHighSpeed.2 = Gauge32: 10000

Fixes: https://github.com/net-snmp/net-snmp/pull/409
[ bvanassche: reformatted commit message / combined four patches into one ]