configure.d: fix compatibility with Clang 16
Clang 16 makes -Wimplicit-function-declaration and -Wimplicit-int an error by
default. Unfortunately, this can lead to misconfiguration or miscompilation of
software as configure tests may then return the wrong result.

We also fix -Wstrict-prototypes while here as it's easy to do and it prepares
us for C23.

For more information, see LWN.net [0] or LLVM's Discourse [1], the Gentoo wiki
[2], or the (new) c-std-porting mailing list [3].

[0] https://lwn.net/Articles/913505/
[1] https://discourse.llvm.org/t/configure-script-breakage-with-the-new-werror-implicit-function-declaration/65213
[2] https://wiki.gentoo.org/wiki/Modern_C_porting
[3] hosted at lists.linux.dev.

Bug: https://bugs.gentoo.org/879715
Signed-off-by: Sam James <sam@gentoo.org>
[ bvanassche: removed two header file guards ]