configure: Fix false positive when build dir contains "yes"
The checks for various IPv6 libraries succeed when the word "yes"
appears anywhere in the CPP output. That's prone to false positives,
e.g. when the build environment path contains that string.

Make the check more robust by anchoring the pattern so the word must
appear on a line of its own.