Spelling: Change 'demon' into 'daemon'
From https://en.wikipedia.org/wiki/Daemon_(computing):

Many people equate the word "daemon" with the word "demon", implying some
kind of satanic connection between UNIX and the underworld. This is an
egregious misunderstanding. "Daemon" is actually a much older form of
"demon"; daemons have no particular bias towards good or evil, but rather
serve to help define a person's character or personality. The ancient
Greeks' concept of a "personal daemon" was similar to the modern concept of
a "guardian angel"-eudaemonia is the state of being helped or protected by a
kindly spirit. As a rule, UNIX systems seem to be infested with both daemons
and demons.