Set flags in MakefileSubs.pm at configure time
Setting the flags at configure time instead of runtime allows use of the
perl module without pulling in net-snmp-config (which in the Fedora/RHEL
RPM setup also pulls in gcc and *-devel packages, not otherwise needed
when just using the perl SNMP module).  The in-tree module is updated
with the rest of the net-snmp code anyway, so setting this at configure
time should stay consistent.

Signed-off-by: Chris Adams <linux@cmadams.net>