agent: release memory back to the system after freeing cache data
The heap memory increases, when retrieving IP-FORWARD MIB objects, in
proportion to the number of entries in the routing table.
It is not given back to the system when freed due to a glibc limitation
(See the link below).

The IP-FORWARD MIB implementation as well as other MIB implementations
use data caching.
Cache management depends on cache flags. For instance, based on
NETSNMP_CACHE_AUTO_RELOAD flag, some caches are reloaded automatically,
while others are reloaded on request.
Each cache has also a timeout. The timeout is checked on cache access.
If expired, the cache is reloaded.
A garbage collector also checks the timeout expiry every minute to
release cache resources.

Call malloc_trim() to return freed memory to the system after the
garbage collector has freed cache data only if the cache will not be
automatically reloaded.

In the best-case scenario, malloc_trim will only be called once after
cache accesses have been completed.

Worst-case, the GC detects the timeout expiry just before a request.
Hence, there will be a wasteful trim, as the cache will be immediately
loaded later. There is a way to avoid this behavior by introducing
NETSNMP_CACHE_RESET_TIMER_ON_USE flag which will reset cache timeout
on each cache access but this may lead to a stale cache if the polling
happens too fast.

Avoid introducing this flag since it depends on polling period.

Note: malloc_trim has a linear complexity, it seems to be safe but it may
softly degrade performance when frequently called.

Link: https://sourceware.org/bugzilla/show_bug.cgi?id=15321

Signed-off-by: Ilyes Ben Hamouda <ilyes.ben_hamouda@6wind.com>