apps/snmptrapd_handlers: Remove a superfluous pointer check
Because the third argument passed to command_handler() is never NULL, remove
a NULL check for that argument. This patch addresses the following Coverity
report:

CID 399166 (#1 of 1): Dereference before null check (REVERSE_INULL)
check_after_deref: Null-checking handler suggests that it may be null, but it
has already been dereferenced on all paths leading to the check.