Fix regression of memory leak when using RPMDB macros.
This change is fixing memory leak when using RPMDB macros. The fix was
introduced in PR #49 but was overwritten over time. Fix is needed in
agent/mibgroup/host/hr_swinst.c only as the regression in the
agent/mibgroup/host/data_access/swinst_rpm.c was fixed few weeks ago.