SIGHUP for handling agent port change
Fix - When the reconfig flag is set, we "shudown master agent" which closes
the socket listening on 161 and "init_master_agent" would establish the
socket on new port without changing the PID of "snmpd".

Use-case where its failing - Though this change works for most of the
cases, its failing in a peculiar case where when we send SIGHUP while
parallely the snmpwalk is in progress. When we send SIGHUP in this usecase,
the socket is closed by "shutdown_master_agent" but a new socket is never
re-established by "init_master_agent" no matter how many sighups we sent,
the snmp-server is down permanently.
Going to the root of the problem, I see that the bind is successful as kernel
returns 0 but netstat(-nlpu) never reports that socket.