Update man page for net-snmp-create-v3-user
Missing option for authentication algorithm and missing parameters for encryption algorithms have been documented.

Although case does not matter for `-A` and `-X` (which seems to be its own can of worms that needs to be kept for compatibility), I've switched them around to match the usage string of `net-snmp-create-v3-user --help`.